resume_job_matcher/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.0
python-multipart==0.0.20
httpx==0.28.1
numpy==1.26.4

# Resume parsing
pypdf==5.1.0
python-docx==1.1.2

# Gemini REST client uses httpx; no official SDK required

# Vector DB (FAISS is most reliable on Linux; Windows users may need conda)
faiss-cpu==1.9.0.post1; platform_system != "Windows"

# PDF generation
reportlab==4.0.9


